refactoring auth
